chore(scap): move machine_info to scap_platform
