+++ userspace/libscap/CMakeLists.txt
@@ -63,7 +63,6 @@ list(APPEND targetfiles
 	scap_platform_api.c
 	scap_procs.c
 	scap_userlist.c
-	scap_machine_agent.c
 	scap_suppress.c)
 
 if(CMAKE_SYSTEM_NAME MATCHES "Linux")
