+++ userspace/libscap/engine/savefile/scap_savefile.c
@@ -2042,6 +2042,8 @@ void scap_savefile_fseek(struct scap_engine_handle engine, uint64_t off)
 
 static int32_t scap_savefile_early_init_platform(struct scap_platform* platform, char* lasterr, struct scap_open_args* oargs)
 {
+	platform->m_machine_info.num_cpus = (uint32_t)-1;
+
     return SCAP_SUCCESS;
 }
 
@@ -2157,7 +2159,7 @@ static int32_t init(struct scap* main_handle, struct scap_open_args* oargs)
 	res = scap_read_init(
 		handle,
 		reader,
-		&main_handle->m_machine_info,
+		&platform->m_machine_info,
 		&platform->m_proclist,
 		&platform->m_addrlist,
 		&platform->m_userlist,
@@ -2225,7 +2227,7 @@ static int32_t scap_savefile_restart_capture(scap_t* handle)
 	if((res = scap_read_init(
 		engine,
 		engine->m_reader,
-		&handle->m_machine_info,
+		&platform->m_machine_info,
 		&platform->m_proclist,
 		&platform->m_addrlist,
 		&platform->m_userlist,
